# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dance Platform API",
    "description": "Complete API documentation for the Dance Platform application. This API provides endpoints for managing users, classes, events, bookings, and more.",
    "version": "2.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@danceplatform.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.danceplatform.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth-token",
        "description": "Authentication cookie"
      }
    },
    "schemas": {
...
      "Class": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Hip Hop Fundamentals"
          },
          "description": {
            "type": "string",
            "example": "Learn the basics of hip hop dance"
          },
          "level": {
            "type": "string",
            "example": "Beginner"
          },
          "durationMins": {
            "type": "integer",
            "example": 60
          },
          "maxCapacity": {
            "type": "integer",
            "example": 30
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "example": 25.00
          },
          "scheduleDays": {
            "type": "string",
            "example": "Monday, Wednesday, Friday"
          },
          "scheduleTime": {
            "type": "string",
            "example": "6:00 PM"
          },
          "requirements": {
            "type": "string",
            "example": "Comfortable clothing, sneakers"
          },
          "imageUrl": {
            "type": "string",
            "example": "/images/hiphop-class.jpg"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Summer Dance Workshop"
          },
          "description": {
            "type": "string"
          },
          "eventType": {
            "type": "string",
            "example": "Workshop"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "example": "17:00"
          },
          "venueId": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "decimal"
          },
          "maxAttendees": {
            "type": "integer"
          },
          "currentAttendees": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PUBLISHED", "CANCELLED"]
          },
          "isFeatured": {
            "type": "boolean"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "An error occurred"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 100
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
    },
    "parameters": {
      "pageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        }
      },
      "limitParam": {
        "name": "limit",
        "in": "query",
        "description": "Number of items per page",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        }
      },
      "searchParam": {
        "name": "search",
        "in": "query",
        "description": "Search term",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Authentication required"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Admin privileges required"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/admin/classes": {
      "get": {
        "tags": ["Admin - Classes"],
        "summary": "List all classes",
        "description": "Retrieve a paginated list of all classes",
        "security": [
          {"bearerAuth": []},
          {"cookieAuth": []}
        ],
        "parameters": [
          {"$ref": "#/components/parameters/pageParam"},
          {"$ref": "#/components/parameters/limitParam"},
          {"$ref": "#/components/parameters/searchParam"},
          {
            "name": "level",
            "in": "query",
            "description": "Filter by class level",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Classes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "classes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Class"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "post": {
        "tags": ["Admin - Classes"],
        "summary": "Create new class",
        "description": "Create a new dance class",
        "security": [
          {"bearerAuth": []},
          {"cookieAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "level", "durationMins", "maxCapacity", "price"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "level": {
                    "type": "string"
                  },
                  "durationMins": {
                    "type": "string"
                  },
                  "maxCapacity": {
                    "type": "string"
                  },
                  "price": {
                    "type": "string"
                  },
                  "scheduleDays": {
                    "type": "string"
                  },
                  "scheduleTime": {
                    "type": "string"
                  },
                  "requirements": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "instructorIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "styleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Class created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "class": {
                      "$ref": "#/components/schemas/Class"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
    "/api/admin/events": {
      "get": {
        "tags": ["Admin - Events"],
        "summary": "List all events",
        "description": "Retrieve a paginated list of all events",
        "security": [
          {"bearerAuth": []},
          {"cookieAuth": []}
        ],
        "parameters": [
          {"$ref": "#/components/parameters/pageParam"},
          {"$ref": "#/components/parameters/limitParam"},
          {"$ref": "#/components/parameters/searchParam"},
          {
            "name": "status",
            "in": "query",
            "description": "Filter by event status",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "PUBLISHED", "CANCELLED"]
            }
          },
          {
            "name": "venueId",
            "in": "query",
            "description": "Filter by venue ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Event"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "post": {
        "tags": ["Admin - Events"],
        "summary": "Create new event",
        "description": "Create a new event",
        "security": [
          {"bearerAuth": []},
          {"cookieAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "eventType", "startDate", "endDate", "venueId", "price", "maxAttendees"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "eventType": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "startTime": {
                    "type": "string"
                  },
                  "endTime": {
                    "type": "string"
                  },
                  "venueId": {
                    "type": "string"
                  },
                  "price": {
                    "type": "string"
                  },
                  "maxAttendees": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["DRAFT", "PUBLISHED", "CANCELLED"]
                  },
                  "isFeatured": {
                    "type": "boolean"
                  },
                  "styleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "event": {
                      "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
  }
}